requirements.txt
flask>=2.0
flask-cors>=3.0
pandas>=1.3
python-Levenshtein>=0.12
unidecode>=1.3.4
indic-transliteration>=2.1